prediction/whisperx/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu121
torch==2.7.1
torchaudio==2.7.1
whisperx==3.4.2
fastapi
uvicorn
python-multipart
google-cloud-storage